requirements.txt
azure-functions
azure-identity
azure-keyvault-secrets
python-tds
pandas
xlsxwriter
python-dateutil
